+++ release/ray_release/byod/build.py
@@ -1,0 +1,11 @@
+from typing import List
+
+from ray_release.test import Test
+
+
+def build_anyscale_byod_images(tests: List[Test]) -> None:
+    """
+    Builds the Anyscale BYOD images for the given tests.
+    """
+    # TODO(aslonnie): Build anyscale byod images for the given ray images
+    return
